extract to config file
